Extract setup logic into a separate file
