+++ components/telemetry-operator/controller/logparser/setup_test.go
@@ -1,0 +1,1 @@
+package logparser
