+++ components/telemetry-operator/controller/logpipeline/controller.go
@@ -19,25 +19,17 @@ package logpipeline
 import (
 	"context"
 	"fmt"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/metrics"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/source"
-
 	telemetryv1alpha1 "github.com/kyma-project/kyma/components/telemetry-operator/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/kyma/components/telemetry-operator/controller"
 	"github.com/kyma-project/kyma/components/telemetry-operator/internal/configchecksum"
 	configbuilder "github.com/kyma-project/kyma/components/telemetry-operator/internal/fluentbit/config/builder"
 	"github.com/prometheus/client_golang/prometheus"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/metrics"
 )
 
 const checksumAnnotationKey = "checksum/logpipeline-config"
@@ -82,74 +74,6 @@ func NewReconciler(client client.Client, config Config, prober DaemonSetProber, 
 	r.syncer = syncer{client, config}
 
 	return &r
-}
-
-func (r *Reconciler) SetupWithManager(mgr ctrl.Manager) error {
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&telemetryv1alpha1.LogPipeline{}).
-		Watches(
-			&source.Kind{Type: &corev1.Secret{}},
-			handler.EnqueueRequestsFromMapFunc(r.mapSecrets),
-			builder.WithPredicates(onlyUpdate()),
-		).
-		Watches(
-			&source.Kind{Type: &appsv1.DaemonSet{}},
-			handler.EnqueueRequestsFromMapFunc(r.mapDaemonSets),
-			builder.WithPredicates(onlyUpdate()),
-		).
-		Complete(r)
-}
-
-func onlyUpdate() predicate.Predicate {
-	return predicate.Funcs{
-		CreateFunc:  func(event event.CreateEvent) bool { return false },
-		DeleteFunc:  func(deleteEvent event.DeleteEvent) bool { return false },
-		UpdateFunc:  func(updateEvent event.UpdateEvent) bool { return true },
-		GenericFunc: func(genericEvent event.GenericEvent) bool { return false },
-	}
-}
-
-func (r *Reconciler) mapSecrets(object client.Object) []reconcile.Request {
-	secret := object.(*corev1.Secret)
-	var pipelines telemetryv1alpha1.LogPipelineList
-	var requests []reconcile.Request
-	err := r.List(context.Background(), &pipelines)
-	if err != nil {
-		ctrl.Log.Error(err, "Secret UpdateEvent: fetching LogPipelineList failed!", err.Error())
-		return requests
-	}
-
-	ctrl.Log.V(1).Info(fmt.Sprintf("Secret UpdateEvent: handling Secret: %s", secret.Name))
-	for i := range pipelines.Items {
-		var pipeline = pipelines.Items[i]
-		if hasSecretRef(&pipeline, secret.Name, secret.Namespace) {
-			request := reconcile.Request{NamespacedName: types.NamespacedName{Name: pipeline.Name}}
-			requests = append(requests, request)
-			ctrl.Log.V(1).Info(fmt.Sprintf("Secret UpdateEvent: added reconcile request for pipeline: %s", pipeline.Name))
-		}
-	}
-	return requests
-}
-
-func (r *Reconciler) mapDaemonSets(object client.Object) []reconcile.Request {
-	daemonSet := object.(*appsv1.DaemonSet)
-
-	var requests []reconcile.Request
-	if daemonSet.Name != r.config.DaemonSet.Name || daemonSet.Namespace != r.config.DaemonSet.Namespace {
-		return requests
-	}
-
-	var allPipelines telemetryv1alpha1.LogPipelineList
-	if err := r.List(context.Background(), &allPipelines); err != nil {
-		ctrl.Log.Error(err, "DaemonSet UpdateEvent: fetching LogPipelineList failed!", err.Error())
-		return requests
-	}
-
-	for _, pipeline := range allPipelines.Items {
-		requests = append(requests, reconcile.Request{NamespacedName: types.NamespacedName{Name: pipeline.Name}})
-	}
-	ctrl.Log.V(1).Info(fmt.Sprintf("DaemonSet changed event handling done: Created %d new reconciliation requests.\n", len(requests)))
-	return requests
 }
 
 func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (reconcileResult ctrl.Result, reconcileErr error) {
