+++ components/telemetry-operator/controller/logpipeline/setup_test.go
@@ -1,0 +1,1 @@
+package logpipeline
