+++ components/telemetry-operator/controller/tracepipeline/setup_test.go
@@ -1,0 +1,1 @@
+package tracepipeline
